Finished JWT
Just need to finish up the request function and then i am done with
authenticating with google(hopefully).
Probably not though - as the only credentials i have accounted for is
the service account itself, i highly doubt it will be possible to
get access to normal google users with just this.