Commit long overdue
Finished an AoC problem.

Continued with the playlist diff project.
Being more familiar with lists and actually knowing what data to
expect from both servers/services, combined with the growing complexity
of making a data structure to combine previous data structures for the
diff search - Has made me conclude that the sanest action as of now, is
to refactor my initial datastructures to make further development easier
and more streamlined.

So a huge refactor is upcomming. :/