Finally statet a bit on the project again.
Had some trouble with geting the correct fonts and stuff working.
Had to make a new flask server because the previous one was in the
hidden_from_git folder.
Reconstructed the hidden files, and documented how to set them up so i
don't have to read through the codebase to remake them the next time my
computer fucks itself over.